Fix nhs registration callback error
When a user attempts to register at the nhs-number
stage in the journey an error occurs because the
app is attempting to access data that does not exist
when persisting the form answers.
The nhs user info is now updated in the session and
form answers and the data persistence is deferred
until later in the journey.