Continuously run a smoke test against the form
At the moment this is a very simple smoke test. We should be do
something smarter but this is a start.

If the smoke test is successful then we curl Cronitor which is
configured as a heartbeat to expect a request every _n_ minutes.

PagerDuty can then be triggered off this.