Implement cookie banner and Google analytics
The cookie banner was not present when a user arrived at the
website for the first time. Google analytics tracking was also
not implemented.
The cookie banner has been implemented in a similar manner to
the GOV UK GovWifi web application.
The Google analytics implementation is similar to the analyticsInit
function that can be found in govuk-coronavirus-vulnerable-people-from
GitHub repository.