Implement pen test security fixes
The penetration test report indicated some fixes are required in
relation to missing security headers and the session cookie not being
marked as secure.

This pull request introduces the talisman package in a similar manner
to the coronavirus-support-service repository to help with implementing
a content security policy, adding security headers and securing the
session cookie. Furthermore, the session is now cleared when an error
occurs.