Exclude non-postable addresses from address picker
The address picker gets address data for postcodes from the Ordnance
Survey Places API. The API supports two datasets, a DPA dataset, which
is compiled by Royal Mail and includes all residential and most
commercial postal addresses; and an LPI dataset, which is compiled by
local authorities and contains both postal addresses, and non-postable
assets such as carparks and streets, which should not appear in the
address picker. We are using the LPI dataset on the address picker.

The address picker should probably use the DBA dataset. However, DPA
records have a different structure to LPI records, which would require a
more extensive rewrite of the postcode_lookup_helper. The documentation
about these datasets is poor, so I've made the decision to simply
exclude records where `POSTAL_ADDRESS_CODE=N`, which indicates a
non-postable address.

This probably needs a more thorough rewrite at some point to make sure
we are using the OS Places API efficiently, and to add a thorough set of
unit tests.

1. https://apidocs.os.uk/docs/os-places-dpa-output
2. https://apidocs.os.uk/docs/os-places-lpi-output
3. https://gbgplc.zendesk.com/hc/en-us/articles/207710315-AddressBase-Premium-Codes