Add tests for the postcode_lookup_helper module
Add some pytest tests for the postcode_lookup_helper, which is the code
that integrates with the OS Places API to populate the address picker
with address info.

The tests use unittest.mock to patch the requests.get() call in the
get_addresses_from_postcode function to use a mocked version that can be
fed fake OS Places API responses.

This also adds new FakeOSPlacesAPIEntry and FakeOSPlacesAPIResponse
classes and corresponding fixtures, which provide a way of generating
fake OS Places API responses, using Faker to generate the fake data. The
structure of the JSON responses is based on real world API responses.