Correct postcode input validation (regex)
The postcode validation regex was missing the start ( ^ )
and end ( $ ) characters. Therefore, it was possible to get
the validation to pass by entering a valid postcode followed
by any random characters.

The input validation erroneously allowed invalid postcodes to be
passed to the postcode eligibility check stored procedure
resulting in an error.