Fix navigation error for returning user
Given a user has signed in using NHS login after
a submission when they view their and answers and
use the 'change' link to modify their address then
an error occurs due to a session value not being
present.

The error has been corrected by using the 'get'
function on the session dict to handle the scenario
where the key is not present.