Update NHS login registration callback logic
When a user attempts to register for an  NHS login account
and a previous submission exists then the user should be
directed to their previously submitted answers.

The NHS registration callback now checks the database for
the existence of a previous submission and redirects the user
to the view-answers page accordingly.