Amend address-lookup back link (NHS login user)
Given a user signs in to their NHS login account and
continues through the journey to the /address-lookup
page, when they click the back link it erroneously
navigates them to the /date-of-birth page.

The assignment of the URL for the back link now takes
into consideration if the user is an NHS login user.
Only non NHS login users are now directed to the
/date-of-birth page.