Add PROD environment (automated) smoke test
A smoke test for the production environment is required
that will not record a submission to the PROD database.

A new behave feature file has been added for the happy
path journey that progresses to the /check-your-answers
page (just prior to persisting answers to the database).

The Makefile and the concourse smoke-test.yml have been
updated accordingly.