847 Investigation of connection sharing
We noticed lots of connections to the database and that we were creating
a new client each time we did a query.

This is an investigation into maintaining a persistent http connection
to aws.

It uses the flask g object to store a connection which has a the headers
set to keep alive.