847 Remove boto connection from app context
There is no explicit close on boto connections (which may make this PR
non-functional, as I have seen something that suggests it is keeping the
connection open between requests anywa, it might be the backend that is
poorly managing the connection to rds?).

So removing it is the best you can do.