Strip spaces & hyphens from nhs number
Users are allowed to submit hyphens which in turn breaks the
stored proc call to RDS as the length of the nhs number is
greater than 10 characters.

The nhs clean function has been utilised to remove whitespace
and hypens from the value submitted in the form.

Additionally, the nhs number validation function checks the
length of the cleansed nhs number now rather than relying
on it's own internal invocation of the clean function.