Fix postcode input validation
The regex for the postcode input validation was
incorrectly allowing partial postcodes through,
e.g. NE11

Two separate input validation regular expressions
are now used to check only alphanumeric characters
are present & to assert the postcode is valid.

The regex for verifying the validity of the postcode
is based on the implementation in the alphagov/collections
repository.