Format postcode before validation (address entry)
The /support-address page was missing a call to the
format_postcode function prior to the invocation of
input validation.

This was causing mis-reporting of input validation
failures when a postcode was entered with a
white space.

The format_postcode function is now invoked in a
consistent manner with the /postcode-eligibility
and /postcode-lookup pages.