Fix AWS RDS secret/db ARN refresh (on error)
An error was occuring in the live environment
when the AWS RDS keys were being rotated.

It transpired that the error was due to
incorrect assignment of the RDS secret/db ARNs
when an error occurs.

A single method has now been created to set the
ARNS in the app configuration.