Pin smoke test chromedriver to 1.1
We have been getting intermittent smoke test timeouts with the
chromedriver as it seems to fail to start.

This is causing alerts to occur. This seems to be a recent change, due
to some change in the gdscyber/chrome-concours-driver. It has gone
forward 5 different versions in the last 8 days, from 1.1 to 1.6.

1.1 is the last version that is significantly different to 1.6 and also
  works with the e2e tests.

It is hard to determine what master was before.