Adds pipeline for sandbox
This creates a sandbox environment for testing changes before they are
merged to master.

It is not to be merged while development is ongoing, it will act as a way
to easily change which branch is pointed to by concourse for this
pipeline.

In order to update the branch that the code is deployed from, change the branch
referred to in the concourse/pipeline-sandbox.yml git-test resource section from
update_concourse_for_sandbox_env to the branch to use, on the
update_concourse_for_sandbox_env

This will keep the pipeline definition on the
update_concourse_for_sandbox_env but change which code gets deployed.