Fix edit name - dob validation error - nhs user
When a user's details are retrieved from the NHS login
API and subsequently used to populate the form_answers()
the data type for day/month/year is an int rather than a
string (the string data type is present when the user
manually enters their dob via the UI).

This was causing the validate_date_of_birth function to
raise an error as it was expecting to receive a string
when verifying the correct quanity of digits have been
entered for the day, month and year respectively.

This error was manifesting when an NHS login user was
attempting to change their name from either the
/check-your-answers (new submission journey) or
/view-answers (returning user journey) pages.

A fix has been implemented to only check the length
of values that are of the type string.