Implement (initial) postcode tier business logic
Postcode eligibility is now determined by using the postcode
to get the LAD (local authority district) code and then using
the LAD code to determine the tier status for the local
authority.

New pages have been added to inform the user that their postcode
is ineligible because either the postcode has not been matched
to a LAD OR the LAD has no tier associated with it.

Some automated test coverage is still required and will be
addressed in imminent/forthcoming ticket(s).