Version smoke and e2e tests based on what has testing for that env
We are about to change the journey quite significantly, putting the
address picker before any other questions.

This means that the behave tests will be out of sink for staging and
production.

In order to make sure the right tests are run against the right
environment, rely on the e2e tests having been run for staging and the
code having been deployed for prod tests.