submission_tracing to output an anonymised event after writing to the db
This introduces a `submission_tracing` integration that has a
`persist_answers_log` method. The aim is to write a stand-alone, detailed but
anonymised, log event to standard out after writing to the database.

The following are taken from the submission:
- submission_reference
- nhs_sub

Additionally, the following are grouped as `submission_details` and shuffled:
- phone_number_calls
- phone_number_texts
- email
- address_line_1
- address_postcode

All of the above goes through a _peppering_ step, the `peppering` method does
the following:
1. takes an object
2. converts it to a string
3. lower cases it
4. removes non-alphanumeric characters
5. SHA256 hashes it with a pepper (`SUBMISSION_TRACING_PEPPER` config item)
6. returns the first 32 characters of the 64 long hex digest

The following are also logged:
- time
- event name (`persist_answers`)
- user agent
- client IP

SUBMISSION_TRACING_PEPPER is provided as a concourse variable (per env) and set
via the concourse pipeline