Address picker ui improvements
This adds in the option "N addresses found" as the first option
to prevent people selecting the default option by mistake.

It also validates that this is not the option people pick.

There was a bug where the validation was not showing up on the page, due
to it populating the wrong error_items. This has been fixed.