Remove padding and add formatting
The padding for the anchor link was putting it in a weird place.

Added presentation role to help with screen readers according to

https://codepen.io/haltersweb/pen/eByVMK

And made the anchor an in-line block to make it conform to the size of
the svg for the tab location.

I tried button elements but they messed with the query parameters.