Moving to a session based backlink
The previous_path is now retrieved in the render page from the session and defaults
to the main govuk start page when it is not set.

It is kept up to date in the blueprint

There is some slight complexity in knowing when a back button has been
pressed or not. It assumes if the path is the same as the previous path
the back button has been pressed it makes an exception for
check-your-answers as that often goes  check-your-answers -> page
-> check-your-answers and otherwise looks like a back button.