Fix problems with sign in with nhs login
This PR has two parts

The first part was to make the svg unfocusable. Some resarch found that
setting it as disabled worked. This was tested using virtual box and the
virtualbox vm, here https://developer.microsoft.com/en-us/microsoft-edge/tools/vms/

The second change sets the role of the anchor element to button. This should fix
the confusion of the dragon users

How this now behaves in nvda, I am unsure.