Log the page source/browser status after failure
Before we were logging the page source/browser status before the test
failure. This change makes it log after the failure. I'm hoping this
will catch any error messages that may have been returned as a result of
flakey `input` interactions and successful form submissions (eg.
Selenium doesn't fill a form correctly, is able to submit and the user
is redirected to the original page, but with errors displayed).