pipeline: fix slack alerts
Currently, when the pipeline fails, we often get *two* slack alerts,
one for "failed" and one for "fixed".  This is because we have
`on_failure` and `on_success` blocks on two different tasks in the
same job.  The `webform-entry` task succeeds, so it emits a "fixed"
notification, then the `s3-validation` task fails, so it emits a
"failed" notification.

The fix is to run all tasks in a single `do` step, and attach the
`on_failure` and `on_success` modifiers to the whole `do` sequence,
not to the individual tasks.